# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Fairview Realty API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  propertyId: string;
  userId: number;
  username: string;
  text: string;
  createdAt: string;
}